+++ src/main/java/sparksoniq/jsoniq/runtime/iterator/functions/sequences/general/HeadFunctionIterator.java
@@ -1,10 +1,7 @@
-package sparksoniq.jsoniq.runtime.iterator.functions.sequences;
+package sparksoniq.jsoniq.runtime.iterator.functions.sequences.general;
 
-import com.sun.javafx.scene.control.ReadOnlyUnbackedObservableList;
 import sparksoniq.exceptions.IteratorFlowException;
-import sparksoniq.jsoniq.item.BooleanItem;
 import sparksoniq.jsoniq.item.Item;
-import sparksoniq.jsoniq.item.metadata.ItemMetadata;
 import sparksoniq.jsoniq.runtime.iterator.RuntimeIterator;
 import sparksoniq.jsoniq.runtime.iterator.functions.base.LocalFunctionCallIterator;
 import sparksoniq.jsoniq.runtime.metadata.IteratorMetadata;
