+++ src/main/java/sparksoniq/jsoniq/runtime/iterator/functions/sequences/general/ReverseFunctionIterator.java
@@ -1,12 +1,7 @@
-package sparksoniq.jsoniq.runtime.iterator.functions.sequences;
+package sparksoniq.jsoniq.runtime.iterator.functions.sequences.general;
 
 import sparksoniq.exceptions.IteratorFlowException;
-import sparksoniq.exceptions.NonAtomicKeyException;
-import sparksoniq.exceptions.UnexpectedTypeException;
-import sparksoniq.jsoniq.item.ArrayItem;
-import sparksoniq.jsoniq.item.IntegerItem;
 import sparksoniq.jsoniq.item.Item;
-import sparksoniq.jsoniq.item.ObjectItem;
 import sparksoniq.jsoniq.runtime.iterator.RuntimeIterator;
 import sparksoniq.jsoniq.runtime.iterator.functions.base.LocalFunctionCallIterator;
 import sparksoniq.jsoniq.runtime.metadata.IteratorMetadata;
@@ -51,7 +46,7 @@ public class ReverseFunctionIterator extends LocalFunctionCallIterator {
 
         List<Item> items = getItemsFromIteratorWithCurrentContext(sequenceIterator);
 
-        for (int i = items.size()-1; i >= 0; i--) {
+        for (int i = items.size() - 1; i >= 0; i--) {
             results.add(items.get(i));
         }
