+++ src/main/java/sparksoniq/jsoniq/runtime/iterator/functions/sequences/general/TailFunctionIterator.java
@@ -1,4 +1,4 @@
-package sparksoniq.jsoniq.runtime.iterator.functions.sequences;
+package sparksoniq.jsoniq.runtime.iterator.functions.sequences.general;
 
 import sparksoniq.exceptions.IteratorFlowException;
 import sparksoniq.jsoniq.item.Item;
@@ -7,7 +7,6 @@ import sparksoniq.jsoniq.runtime.iterator.functions.base.LocalFunctionCallIterat
 import sparksoniq.jsoniq.runtime.metadata.IteratorMetadata;
 import sparksoniq.semantics.DynamicContext;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class TailFunctionIterator extends LocalFunctionCallIterator {
