+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionEmpty1.iq
