+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionInsertBefore6.iq
