+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionInsertBeforeError1.iq
