+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionInsertBeforeError2.iq
