+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionInsertBeforeError4.iq
