+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionInsertBeforeError5.iq
