+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionInsertBeforeError6.iq
