+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionRemove1.iq
