+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionRemove2.iq
