+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionRemoveError1.iq
