+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionRemoveError2.iq
