+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionRemoveError4.iq
