+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionRemoveError5.iq
