+++ src/main/resources/test_files/runtime/FunctionSequence/General/FunctionRemoveError6.iq
